refactor(v3.2.1): 抽 579 行 SVG 图元到 lib/report/svg_primitives
- 19 个 svg_xxx 函数 + 9 个 COLOR_* 常量搬到 lib/report/svg_primitives.py
- assemble_report.py 做 re-export 保持向后兼容
- 行数: 2964 → 2400 行 (-19%)
- 332 tests 全过 · e2e assemble 零影响
